+++ apps/web/client/src/app/projects/import/local/_components/verify-project.tsx
@@ -6,11 +6,11 @@ import { Button } from '@onlook/ui/button';
 import { motion } from 'motion/react';
 import { Icons } from '@onlook/ui/icons';
 import { useEffect, useState } from 'react';
-import { useProjectCreation } from './project-creation-context';
+import { useProjectCreation } from '../_context/project-creation-context';
 import { StepFooter } from './steps';
 import { StepContent } from './steps';
 import { StepHeader } from './steps';
-import type { NextJsProjectValidation, ProcessedFile } from '../../types';
+import type { NextJsProjectValidation, ProcessedFile } from '@/app/projects/types';
 
 export const VerifyProject = () => {
     const { projectData, prevStep, nextStep, isFinalizing } = useProjectCreation();
